replace terminator with xfce4-terminal + tmux

+++ setup/setup-readline.sh
@@ -1,1 +1,4 @@
-ln -s "$dotfiles_dir/readline/.inputrc" "$HOME/"
+#!/bin/bash
+if [ ! -f "$HOME/.inputrc" ]; then
+    ln -s "$dotfiles_dir/readline/.inputrc" "$HOME/"
+fi
